requirements.txt
numpy==1.19.5
torch==1.8.0+cu111
torchvision==0.9.0+cu111
Pillow==8.1.2
